+++ Hangfire.FluentNHibernateStorage/JobQueue/FetchedJob.cs
@@ -2,8 +2,8 @@
 {
     public class FetchedJob
     {
-        public int Id { get; set; }
-        public int JobId { get; set; }
+        public long Id { get; set; }
+        public long JobId { get; set; }
         public string Queue { get; set; }
     }
 }
