+++ Hangfire.FluentNHibernateStorage/JobQueue/IPersistentJobQueueMonitoringApi.cs
@@ -5,8 +5,8 @@ namespace Hangfire.FluentNHibernateStorage.JobQueue
     public interface IPersistentJobQueueMonitoringApi
     {
         IEnumerable<string> GetQueues();
-        IEnumerable<int> GetEnqueuedJobIds(string queue, int from, int perPage);
-        IEnumerable<int> GetFetchedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetEnqueuedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetFetchedJobIds(string queue, int from, int perPage);
         EnqueuedAndFetchedCountDto GetEnqueuedAndFetchedCount(string queue);
     }
 }
